capsules/summarize-text/requirements.txt
openai>=1.12.0
pyyaml>=6.0.1
jsonschema>=4.19.0
